RDKB-6178: Wifi reset is blocked when calling router&wifi reset
Reason for change: In field we observed that reboot is called for both router and wifi when erouter is not reachable.
This is a redundant operation, as we need not have to reset wifi in this case. When this happens, we observed that it
could corrupt wifi configuration if customer has already triggered a wifi reset.

So we should be allowing only router reset when router&wifi is called.
Test Procedure: Verified by follwoing below steps:
1) Set MIB to reset both router and wifi by setting value "1"
2) Check whether wifi is reset or not. Wifi should not be reset
3) Check in /rdklogs/logs/SNMPlog.txt.0 file for logs indicating what operation was performed
4) Reset router and wifi individually by passing values 2 and 3 respectively
5) Verify only corresponding modules are reset in step 4

Change-Id: Ide1107327f8690820a5179bf46269ee51bc52385
Risks: None
Signed-off-by: achelakkode<anoopnair_c@infosys.com>